+++ backend/controllers/purchaseController.js
@@ -9,20 +9,18 @@ const setDb = (database) => {
 
 const getAllPurchases = async (req, res) => {
     try {
-        const purchases = await db.find().toArray();
+        const purchases = await db.find({userId: req.user.uid}).toArray();
         res.status(200).json(purchases);
     } catch (error) {
-        console.error(error);
         res.status(500).json({ message: 'Error al obtener las compras.', error });
     }
 };
 const getPurchaseById = async (req, res) => {
     try {
         const id = req.params.id;
-        const purchase = await db.findOne({ _id: new ObjectId(id) })
+        const purchase = await db.findOne({ _id: new ObjectId(id),userId: req.user.uid })
         res.status(200).json(purchase);
     } catch (error) {
-        console.error(error);
         res.status(500).json({ message: 'Error al obtener la compra', error });
     }
 };
@@ -44,7 +42,17 @@ const createPurchase = async (req, res) => {
         }
         const status = "PENDING";
         const purchase = new Purchase(name, supplier, products, totalAmount, paymentMethod, purchaseDate, status, notes);
-        const result = await db.insertOne(purchase);
+        const result = await db.insertOne({
+            userId: req.user.uid,
+            name,
+            purchaseDate,
+            paymentMethod,
+            notes,
+            totalAmount,
+            status,
+            products,
+            supplier
+    });
         if (!result.insertedId) {
             return res.status(500).json({ message: 'Error al insertar la compra' });
         }
@@ -53,7 +61,6 @@ const createPurchase = async (req, res) => {
             purchase: { _id: result.insertedId, ...purchase }
         });
     } catch (error) {
-        console.error('Error al insertar la compra:', error);
         res.status(500).json({ message: 'Error al insertar la compra', error: error.message });
     }
 };
@@ -75,33 +82,31 @@ const updatePurchase = async (req, res) => {
             data = {...data, totalAmount: totalAmount};
         }
         const result = await db.updateOne(
-            { _id: new ObjectId(id) }, 
+            { _id: new ObjectId(id),userId: req.user.uid }, 
             { $set: data } 
         )
         if (result.matchedCount === 0) {
             return res.status(404).json({ message: "Compra no encontrada" });
         }
 
-        const updatedPurchase = await db.findOne({ _id: new ObjectId(id) });
+        const updatedPurchase = await db.findOne({ _id: new ObjectId(id),userId: req.user.uid });
         res.json(updatedPurchase);
     } catch (error) {
-        console.error("Error al actualizar la compra:", error);
         res.status(500).json({ message: "Error al actualizar la compra", error });
     } 
 };
 
-const deletePurchase = async (req, res) => {
+/*const deletePurchase = async (req, res) => {
     try {
         const purchaseId = req.params.id;
-        const result = await db.deleteOne({ _id: new ObjectId(purchaseId) });
+        const result = await db.deleteOne({ _id: new ObjectId(purchaseId),userId: req.user.uid });
         if (result.deletedCount === 0) {
             return res.status(404).json({ message: "Compra no encontrada" });
         }
         res.json({ message: "Compra eliminado correctamente" });
     } catch (error) {
-        console.error("Error al eliminar la compra:", error);
         res.status(500).json({ message: "Error al eliminar la compra" });
     }
-};
+};*/
 
-module.exports = { setDb, getAllPurchases, getPurchaseById, createPurchase, updatePurchase, deletePurchase };
+module.exports = { setDb, getAllPurchases, getPurchaseById, createPurchase, updatePurchase };
