+++ backend/controllers/saleController.js
@@ -9,20 +9,18 @@ const setDb = (database) => {
 
 const getAllSales = async (req, res) => {
     try {
-        const sales = await db.find().toArray();
+        const sales = await db.find({userId: req.user.uid}).toArray();
         res.status(200).json(sales);
     } catch (error) {
-        console.error(error);
         res.status(500).json({ message: 'Error al obtener las ventas.', error });
     }
 };
 const getSaleById = async (req, res) => {
     try {
         const id = req.params.id;
-        const sale = await db.findOne({ _id: new ObjectId(id) })
+        const sale = await db.findOne({ _id: new ObjectId(id),userId: req.user.uid })
         res.status(200).json(sale);
     } catch (error) {
-        console.error(error);
         res.status(500).json({ message: 'Error al obtener la venta', error });
     }
 };
@@ -44,7 +42,16 @@ const createSale = async (req, res) => {
         }
         const status = "PENDING";
         const sale = new Sale(name, products, totalAmount ,status, paymentMethod, saleDate, notes);
-        const result = await db.insertOne(sale);
+        const result = await db.insertOne({
+            userId: req.user.uid,
+            name,
+            products,
+            totalAmount,
+            paymentMethod,
+            saleDate,
+            status,
+            notes
+        });
         if (!result.insertedId) {
             return res.status(500).json({ message: 'Error al insertar la venta' });
         }
@@ -53,7 +60,6 @@ const createSale = async (req, res) => {
             sale: { _id: result.insertedId, ...sale }
         });
     } catch (error) {
-        console.error('Error al insertar la venta:', error);
         res.status(500).json({ message: 'Error al insertar la venta', error: error.message });
     }
 };
@@ -75,33 +81,31 @@ const updateSale = async (req, res) => {
             data = {...data, totalAmount: totalAmount};
         }
         const result = await db.updateOne(
-            { _id: new ObjectId(id) }, 
+            { _id: new ObjectId(id),userId: req.user.uid }, 
             { $set: data } 
         )
         if (result.matchedCount === 0) {
             return res.status(404).json({ message: "Venta no encontrada" });
         }
 
-        const updatedSale = await db.findOne({ _id: new ObjectId(id) });
+        const updatedSale = await db.findOne({ _id: new ObjectId(id),userId: req.user.uid });
         res.json(updatedSale);
     } catch (error) {
-        console.error("Error al actualizar la venta:", error);
         res.status(500).json({ message: "Error al actualizar la venta", error });
     } 
 };
 
-const deleteSale = async (req, res) => {
+/*const deleteSale = async (req, res) => {
     try {
         const saleId = req.params.id;
-        const result = await db.deleteOne({ _id: new ObjectId(saleId) });
+        const result = await db.deleteOne({ _id: new ObjectId(saleId),userId: req.user.uid });
         if (result.deletedCount === 0) {
             return res.status(404).json({ message: "Venta no encontrada" });
         }
         res.json({ message: "Venta eliminado correctamente" });
     } catch (error) {
-        console.error("Error al eliminar la venta:", error);
         res.status(500).json({ message: "Error al eliminar la venta" });
     }
-};
+};*/
 
-module.exports = { setDb, getAllSales,getSaleById, createSale, updateSale, deleteSale };
+module.exports = { setDb, getAllSales,getSaleById, createSale, updateSale };
